backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
pymongo==4.8.0
python-dotenv==1.0.1
Werkzeug==3.0.3
gunicorn==22.0.0
pika==1.3.2
elasticsearch==8.14.0